Create currying.js
Currying is a transformation of a function with multiple arguments into a sequence of nested functions with a single argument. The currying allows to perform function specialization and composition. Curried functions are higher-order functions which allow us to create specialized versions of original functions.

We can transform the fn(a,b,c) callable into fn(a)(b)(c)